fix: extract PromptGenerator interface to unblock MockK and stabilize tests
- PromptGenerator is now an interface; PromptGeneratorImpl holds the
  LiteRT-LM implementation. TriggerPipelineTest and HabitEditViewModelTest
  mock the interface, which is trivially intercepted by MockK regardless
  of LiteRT-LM's effect on the concrete class's bytecode.
- ServiceModule provides PromptGeneratorImpl bound to PromptGenerator
  so Hilt injection continues to work without modification to call sites.
- PromptGeneratorTest is updated to instantiate PromptGeneratorImpl
  directly and stubs context.filesDir with a real temp directory so
  initialize() can resolve File paths without NPE.
- ModelDownloadWorker switches from setProgress() (suspend, awaits a
  ListenableFuture that the relaxed mock never completes) to
  setProgressAsync() (fire-and-forget), eliminating the
  UncompletedCoroutinesError in the atomically-writes test.
